tests/wpt/web-platform-tests/css/css-cascade/at-scope-parsing.html

<!doctype html>
<title>@scope: parsing</title>
<link rel="help" href="https://drafts.csswg.org/css-cascade-6/#scope-atrule">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<main id=main></main>
<script>
  function test_valid(actual, expected) {
    if (expected === undefined)
      expected = actual;
    test(t => {
      t.add_cleanup(() => main.replaceChildren());
      let style = document.createElement('style');
      style.textContent = `${actual}{}`;
      main.append(style);
      assert_equals(style.sheet.rules.length, 1);
      let rule = style.sheet.rules[0];
      assert_equals(rule.cssText, `${expected} {\n}`);
    }, `${actual} is valid`);
  }

  function test_invalid(actual) {
    test(t => {
      t.add_cleanup(() => main.replaceChildren());
      let style = document.createElement('style');
      style.textContent = `${actual}{}`;
      main.append(style);
      assert_equals(style.sheet.rules.length, 0);
    }, `${actual} is not valid`);
  }

  test_valid('@scope (.a)');
  test_valid('@scope (.a + .b)');
  test_valid('@scope (.a:hover)');
  test_valid('@scope (.a:hover, #b, div)');
  test_valid('@scope (:is(div, span))');

  test_valid('@scope (.a) to (.b)');
  test_valid('@scope (.a)to (.b)', '@scope (.a) to (.b)');
  test_valid('@scope (.a) to (.b:hover, #c, div)');
  test_valid('@scope');
  test_valid('@scope to (.a)');
  test_valid('@scope (.a) to (&)');
  test_valid('@scope (.a) to (& > &)');
  test_valid('@scope (.a) to (> .b)');
  test_valid('@scope (.a) to (+ .b)');
  test_valid('@scope (.a) to (~ .b)');

  // Forgiving behavior:
  test_valid('@scope (.c <> .d)', '@scope ()');
  test_valid('@scope (.a, .c <> .d)', '@scope (.a)');
  test_valid('@scope (.a <> .b, .c)', '@scope (.c)');
  test_valid('@scope (div::before)', '@scope ()');
  test_valid('@scope (div::after)', '@scope ()');
  test_valid('@scope (slotted(div))', '@scope ()');
  test_valid('@scope (.a) to (div::before)', '@scope (.a)');

  test_invalid('@scope (.a) unknown (.c)');
  test_invalid('@scope (.a) to unknown (.c)');
  test_invalid('@scope (.a) 1px (.c)');
  test_invalid('@scope (.a) to unknown(c)');
  test_invalid('@scope unknown(.a)');
  test_invalid('@scope 1px');
  test_invalid('@scope creep');
</script>
